- Create a new component Home.js in the components folder and move all the code from App.js to the new component.     - Update the paths and component’s name accordingly
